refactor(conv): rename the output format writer module from "*_parser" to "*_out_writer"
