+++ src/conv/mod.rs
@@ -1,11 +1,7 @@
 pub(crate) mod binary_types;
-pub(crate) mod text_parser;
 pub(crate) mod binary_parser;
-pub(crate) mod xyz_parser;
-pub(crate) mod dump_parser;
-pub(crate) mod bin_parser;
 mod binary_parser_v1;
 mod binary_parser_v2;
 mod v2_atom_types;
 mod lib_conv_capi;
-mod out_writer;
+pub(crate) mod writers;
